Create initial implementation for finding all functions in a file
The idea is that this program will loop over all files in a directory
and match files that are of type *.cpp *.py or whatever has been
previously defined. It will then load the file into a string and find
the defined functions. This information will then be used to extract the
docstring connected to said function. This commit only includes the
discovery of functions in a file.